Turn on clippy pedantic + nursery, pin lints in Cargo.toml, add pre-commit hooks (#24)
* chore: apply clippy pedantic and nursery auto-fixes

Mechanical pass from `cargo clippy --fix` with clippy::pedantic and
clippy::nursery enabled. Clears 269 warnings, mostly Self in place of the
repeated type name, #[must_use] on public getters, const fn, and
map_or/map_or_else in place of map().unwrap_or().

No behaviour change. The one hand edit is init.rs, where the auto-fix
turned a match into if let/else; let...else reads better.


* chore: hand-fix the clippy warnings the auto-fix could not

Docs: # Errors on the seven fallible public functions, # Panics on the
four that can abort, and the four over-long first doc paragraphs split
into a summary line plus a body.

Code: writeln! in place of push_str(&format!(..)) in the renderer, any()
in place of a collect that was only tested for emptiness, let...else in
apply, cast_signed/cast_unsigned in clock, and a handful of
map_or/unwrap_or_else swaps.


* chore: pin clippy pedantic and nursery as deny in the workspace

Lints now live in Cargo.toml rather than only in a CI command line, so a
bare `cargo clippy` gives the same answer locally as CI does. Both crates
opt in via [lints] workspace = true.

The eleven remaining warnings are deliberate and carry #[expect] with a
reason: CLI/config structs that mirror one bool per flag, duration casts
in the min-age checks, uv's write target being None for a different
reason than the legacy Python three, and scan::run holding a non-Send
reporter across an await while never being spawned.

option_if_let_else is allowed workspace-wide: it rewrites the settings
checks' Option matches, which read as tables, into worse closure form.

Also drops anyhow, which was declared but never imported, and points
`repository` at jackmcpickle/pkguard instead of the old
security-audit-tool URL, which both crates would otherwise publish.


* chore: add prek pre-commit hooks for fmt and clippy

cargo fmt rewrites in place; when it does, prek blocks the commit so the
rewrite gets re-staged deliberately. Clippy never rewrites — it blocks
and leaves the fix to a human, because clippy --fix makes semantic
changes that deserve a read first.

CI now runs the same clippy invocation as the hook. Levels come from the
workspace [lints] table, so `cargo clippy` with no flags, the hook, and
CI all agree.

Setup is one line, documented in AGENTS.md:
  brew install prek && prek install


* chore: remove old ts skill

* fix(ci): satisfy map_unwrap_or on Result, new in clippy 1.98

CI runs the latest stable toolchain, which had moved to 1.98 while this
branch was developed against 1.93. clippy 1.98 extends map_unwrap_or to
Result, flagging four sites the older clippy passed.

Also adds .github/dependabot.yml with one grouped weekly PR per
ecosystem: cargo, npm for the pnpm site, and github-actions. Version
updates group as minor+patch; security updates group separately so they
are not held behind a routine bump.

Note for future edits: there is no `pnpm` package-ecosystem value.
Dependabot covers npm, Yarn, and pnpm under `npm`.


* fix(ci): regenerate docs catalog for 1.1.0

The 1.1.0 release commit bumped the workspace version but did not
re-dump site/src/generated/catalog.json, which still carried 1.0.3. CI
did not run on that commit, so the staleness went unnoticed until this
branch built against 1.1.0.


---------

+++ crates/pkguard-core/src/settings/checks/provenance.rs
@@ -5,7 +5,9 @@ use crate::format::yaml::{self, Yaml};
 use crate::manager::{Manager, PackageManagerPin};
 use std::path::Path;
 
-const TRUST_POLICY_IGNORE_AFTER_MINUTES: f64 = 90.0 * 24.0 * 60.0;
+/// 90 days. Held as an integer so both the comparison and the written value
+/// derive from it losslessly.
+const TRUST_POLICY_IGNORE_AFTER_MINUTES: i32 = 90 * 24 * 60;
 
 pub fn pnpm_checks(yaml: &Yaml, yaml_path: &Path, pin: Option<&PackageManagerPin>) -> Vec<Finding> {
     let mut findings = Vec::new();
@@ -15,7 +17,7 @@ pub fn pnpm_checks(yaml: &Yaml, yaml_path: &Path, pin: Option<&PackageManagerPin
             &["trustPolicyIgnoreAfter", "trust-policy-ignore-after"],
         )
         .and_then(yaml::as_f64);
-        if !minutes.is_some_and(|m| m >= TRUST_POLICY_IGNORE_AFTER_MINUTES) {
+        if !minutes.is_some_and(|m| m >= f64::from(TRUST_POLICY_IGNORE_AFTER_MINUTES)) {
             findings.push(fixable_finding(
                 "provenance.ignore-after",
                 format!(
@@ -28,7 +30,7 @@ pub fn pnpm_checks(yaml: &Yaml, yaml_path: &Path, pin: Option<&PackageManagerPin
                     yaml_path,
                     vec![ConfigEdit::set(
                         "trustPolicyIgnoreAfter",
-                        TRUST_POLICY_IGNORE_AFTER_MINUTES as i64,
+                        i64::from(TRUST_POLICY_IGNORE_AFTER_MINUTES),
                     )],
                 ),
             ));
